requirements.txt
Flask==2.1.2
firebase-admin
pyrebase4
Werkzeug==2.3.7
Pillow~=10.1.0
transformers~=4.35.2
torch
torchvision
torchaudio
easyocr~=1.7.1
numpy~=1.26.2
openai~=1.3.5
deepface~=0.0.79